+++ proxy/route.go
@@ -6,10 +6,11 @@ import (
 
 // Route describes one way that a request may be mapped from the original HTTP request to an upstream server.
 type Route struct {
-	Domains   []string
-	Upstreams []Upstream
-	Headers   []Header
-	Provider  string
+	Domains           []string
+	Upstreams         []Upstream
+	Headers           []Header
+	Provider          string
+	RedirectToPrimary bool
 
 	certificate       *tls.Certificate
 	certificateStatus CertificateStatus